+++ tableauserverclient/models/datasource_item.py
@@ -1,9 +1,9 @@
 import xml.etree.ElementTree as ET
 from .exceptions import UnpopulatedPropertyError
-from .property_decorators import property_not_nullable
+from .property_decorators import property_not_nullable, property_is_boolean
 from .tag_item import TagItem
-from .. import NAMESPACE
 from ..datetime_helpers import parse_datetime
+import copy
 
 
 class DatasourceItem(object):
@@ -11,14 +11,17 @@ class DatasourceItem(object):
         self._connections = None
         self._content_url = None
         self._created_at = None
-        self._id = None
-        self._project_name = None
-        self._tags = set()
         self._datasource_type = None
+        self._id = None
+        self._initial_tags = set()
+        self._project_name = None
         self._updated_at = None
+        self._certified = None
+        self._certification_note = None
         self.name = name
         self.owner_id = None
         self.project_id = project_id
+        self.tags = set()
 
     @property
     def connections(self):
@@ -34,6 +37,24 @@ class DatasourceItem(object):
     @property
     def created_at(self):
         return self._created_at
+
+    @property
+    def certified(self):
+        return self._certified
+
+    @certified.setter
+    @property_not_nullable
+    @property_is_boolean
+    def certified(self, value):
+        self._certified = value
+
+    @property
+    def certification_note(self):
+        return self._certification_note
+
+    @certification_note.setter
+    def certification_note(self, value):
+        self._certification_note = value
 
     @property
     def id(self):
@@ -53,10 +74,6 @@ class DatasourceItem(object):
         return self._project_name
 
     @property
-    def tags(self):
-        return self._tags
-
-    @property
     def datasource_type(self):
         return self._datasource_type
 
@@ -67,16 +84,18 @@ class DatasourceItem(object):
     def _set_connections(self, connections):
         self._connections = connections
 
-    def _parse_common_tags(self, datasource_xml):
+    def _parse_common_elements(self, datasource_xml, ns):
         if not isinstance(datasource_xml, ET.Element):
-            datasource_xml = ET.fromstring(datasource_xml).find('.//t:datasource', namespaces=NAMESPACE)
+            datasource_xml = ET.fromstring(datasource_xml).find('.//t:datasource', namespaces=ns)
         if datasource_xml is not None:
-            (_, _, _, _, _, updated_at, _, project_id, project_name, owner_id) = self._parse_element(datasource_xml)
-            self._set_values(None, None, None, None, None, updated_at, None, project_id, project_name, owner_id)
+            (_, _, _, _, _, updated_at, _, project_id, project_name, owner_id,
+             certified, certification_note) = self._parse_element(datasource_xml, ns)
+            self._set_values(None, None, None, None, None, updated_at, None, project_id,
+                             project_name, owner_id, certified, certification_note)
         return self
 
     def _set_values(self, id, name, datasource_type, content_url, created_at,
-                    updated_at, tags, project_id, project_name, owner_id):
+                    updated_at, tags, project_id, project_name, owner_id, certified, certification_note):
         if id is not None:
             self._id = id
         if name:
@@ -90,53 +109,61 @@ class DatasourceItem(object):
         if updated_at:
             self._updated_at = updated_at
         if tags:
-            self._tags = tags
+            self.tags = tags
+            self._initial_tags = copy.copy(tags)
         if project_id:
             self.project_id = project_id
         if project_name:
             self._project_name = project_name
         if owner_id:
             self.owner_id = owner_id
+        if certification_note:
+            self.certification_note = certification_note
+        self.certified = certified  # Always True/False, not conditional
 
     @classmethod
-    def from_response(cls, resp):
+    def from_response(cls, resp, ns):
         all_datasource_items = list()
         parsed_response = ET.fromstring(resp)
-        all_datasource_xml = parsed_response.findall('.//t:datasource', namespaces=NAMESPACE)
+        all_datasource_xml = parsed_response.findall('.//t:datasource', namespaces=ns)
 
         for datasource_xml in all_datasource_xml:
-            (id, name, datasource_type, content_url, created_at, updated_at,
-             tags, project_id, project_name, owner_id) = cls._parse_element(datasource_xml)
+            (id_, name, datasource_type, content_url, created_at, updated_at,
+             tags, project_id, project_name, owner_id,
+             certified, certification_note) = cls._parse_element(datasource_xml, ns)
             datasource_item = cls(project_id)
-            datasource_item._set_values(id, name, datasource_type, content_url, created_at, updated_at,
-                                        tags, None, project_name, owner_id)
+            datasource_item._set_values(id_, name, datasource_type, content_url, created_at, updated_at,
+                                        tags, None, project_name, owner_id, certified, certification_note)
             all_datasource_items.append(datasource_item)
         return all_datasource_items
 
     @staticmethod
-    def _parse_element(datasource_xml):
-        id = datasource_xml.get('id', None)
+    def _parse_element(datasource_xml, ns):
+        id_ = datasource_xml.get('id', None)
         name = datasource_xml.get('name', None)
         datasource_type = datasource_xml.get('type', None)
         content_url = datasource_xml.get('contentUrl', None)
         created_at = parse_datetime(datasource_xml.get('createdAt', None))
         updated_at = parse_datetime(datasource_xml.get('updatedAt', None))
+        certification_note = datasource_xml.get('certificationNote', None)
+        certified = str(datasource_xml.get('isCertified', None)).lower() == 'true'
 
         tags = None
-        tags_elem = datasource_xml.find('.//t:tags', namespaces=NAMESPACE)
+        tags_elem = datasource_xml.find('.//t:tags', namespaces=ns)
         if tags_elem is not None:
-            tags = TagItem.from_xml_element(tags_elem)
+            tags = TagItem.from_xml_element(tags_elem, ns)
 
         project_id = None
         project_name = None
-        project_elem = datasource_xml.find('.//t:project', namespaces=NAMESPACE)
+        project_elem = datasource_xml.find('.//t:project', namespaces=ns)
         if project_elem is not None:
             project_id = project_elem.get('id', None)
             project_name = project_elem.get('name', None)
 
         owner_id = None
-        owner_elem = datasource_xml.find('.//t:owner', namespaces=NAMESPACE)
+        owner_elem = datasource_xml.find('.//t:owner', namespaces=ns)
         if owner_elem is not None:
             owner_id = owner_elem.get('id', None)
 
-        return id, name, datasource_type, content_url, created_at, updated_at, tags, project_id, project_name, owner_id
+        return (id_, name, datasource_type, content_url, created_at, updated_at, tags, project_id,
+                project_name, owner_id, certified, certification_note)