+++ tableauserverclient/server/__init__.py
@@ -1,10 +1,10 @@
 from .request_factory import RequestFactory
-from .request_options import RequestOptions
+from .request_options import ImageRequestOptions, RequestOptions
 from .filter import Filter
 from .sort import Sort
 from .. import ConnectionItem, DatasourceItem,\
     GroupItem, PaginationItem, ProjectItem, ScheduleItem, SiteItem, TableauAuth,\
-    UserItem, ViewItem, WorkbookItem, NAMESPACE
+    UserItem, ViewItem, WorkbookItem, TaskItem
 from .endpoint import Auth, Datasources, Endpoint, Groups, Projects, Schedules, \
     Sites, Users, Views, Workbooks, ServerResponseError, MissingRequiredFieldError
 from .server import Server