+++ tableauserverclient/server/endpoint/datasources_endpoint.py
@@ -1,7 +1,10 @@
 from .endpoint import Endpoint, api, parameter_added_in
 from .exceptions import MissingRequiredFieldError
 from .fileuploads_endpoint import Fileuploads
+from .resource_tagger import _ResourceTagger
 from .. import RequestFactory, DatasourceItem, PaginationItem, ConnectionItem
+from ...filesys_helpers import to_filename
+from ...models.tag_item import TagItem
 import os
 import logging
 import copy
@@ -17,6 +20,10 @@ logger = logging.getLogger('tableau.endpoint.datasources')
 
 
 class Datasources(Endpoint):
+    def __init__(self, parent_srv):
+        super(Datasources, self).__init__(parent_srv)
+        self._resource_tagger = _ResourceTagger(parent_srv)
+
     @property
     def baseurl(self):
         return "{0}/sites/{1}/datasources".format(self.parent_srv.baseurl, self.parent_srv.site_id)
@@ -27,8 +34,8 @@ class Datasources(Endpoint):
         logger.info('Querying all datasources on site')
         url = self.baseurl
         server_response = self.get_request(url, req_options)
-        pagination_item = PaginationItem.from_response(server_response.content)
-        all_datasource_items = DatasourceItem.from_response(server_response.content)
+        pagination_item = PaginationItem.from_response(server_response.content, self.parent_srv.namespace)
+        all_datasource_items = DatasourceItem.from_response(server_response.content, self.parent_srv.namespace)
         return all_datasource_items, pagination_item
 
     # Get 1 datasource by id
@@ -40,7 +47,7 @@ class Datasources(Endpoint):
         logger.info('Querying single datasource (ID: {0})'.format(datasource_id))
         url = "{0}/{1}".format(self.baseurl, datasource_id)
         server_response = self.get_request(url)
-        return DatasourceItem.from_response(server_response.content)[0]
+        return DatasourceItem.from_response(server_response.content, self.parent_srv.namespace)[0]
 
     # Populate datasource item's connections
     @api(version="2.0")
@@ -66,18 +73,24 @@ class Datasources(Endpoint):
     # Download 1 datasource by id
     @api(version="2.0")
     @parameter_added_in(no_extract='2.5')
-    def download(self, datasource_id, filepath=None, no_extract=False):
+    @parameter_added_in(include_extract='2.5')
+    def download(self, datasource_id, filepath=None, include_extract=True, no_extract=None):
         if not datasource_id:
             error = "Datasource ID undefined."
             raise ValueError(error)
         url = "{0}/{1}/content".format(self.baseurl, datasource_id)
 
-        if no_extract:
+        if no_extract is False or no_extract is True:
+            import warnings
+            warnings.warn('no_extract is deprecated, use include_extract instead.', DeprecationWarning)
+            include_extract = not no_extract
+
+        if not include_extract:
             url += "?includeExtract=False"
 
         with closing(self.get_request(url, parameters={'stream': True})) as server_response:
             _, params = cgi.parse_header(server_response.headers['Content-Disposition'])
-            filename = os.path.basename(params['filename'])
+            filename = to_filename(os.path.basename(params['filename']))
             if filepath is None:
                 filepath = filename
             elif os.path.isdir(filepath):
@@ -96,12 +109,16 @@ class Datasources(Endpoint):
         if not datasource_item.id:
             error = 'Datasource item missing ID. Datasource must be retrieved from server first.'
             raise MissingRequiredFieldError(error)
+
+        self._resource_tagger.update_tags(self.baseurl, datasource_item)
+
+        # Update the datasource itself
         url = "{0}/{1}".format(self.baseurl, datasource_item.id)
         update_req = RequestFactory.Datasource.update_req(datasource_item)
         server_response = self.put_request(url, update_req)
         logger.info('Updated datasource item (ID: {0})'.format(datasource_item.id))
         updated_datasource = copy.copy(datasource_item)
-        return updated_datasource._parse_common_tags(server_response.content)
+        return updated_datasource._parse_common_elements(server_response.content, self.parent_srv.namespace)
 
     # Publish datasource
     @api(version="2.0")
@@ -144,6 +161,6 @@ class Datasources(Endpoint):
                                                                               file_contents,
                                                                               connection_credentials)
         server_response = self.post_request(url, xml_request, content_type)
-        new_datasource = DatasourceItem.from_response(server_response.content)[0]
+        new_datasource = DatasourceItem.from_response(server_response.content, self.parent_srv.namespace)[0]
         logger.info('Published {0} (ID: {1})'.format(filename, new_datasource.id))
         return new_datasource