+++ tableauserverclient/server/request_factory.py
@@ -1,5 +1,6 @@
 from ..datetime_helpers import format_datetime
 import xml.etree.ElementTree as ET
+from functools import wraps
 
 from requests.packages.urllib3.fields import RequestField
 from requests.packages.urllib3.filepost import encode_multipart_formdata
@@ -14,6 +15,14 @@ def _add_multipart(parts):
     xml_request, content_type = encode_multipart_formdata(mime_multipart_parts)
     content_type = ''.join(('multipart/mixed',) + content_type.partition(';')[1:])
     return xml_request, content_type
+
+
+def _tsrequest_wrapped(func):
+    def wrapper(self, *args, **kwargs):
+        xml_request = ET.Element('tsRequest')
+        func(xml_request, *args, **kwargs)
+        return ET.tostring(xml_request)
+    return wrapper
 
 
 class AuthRequest(object):
@@ -56,6 +65,12 @@ class DatasourceRequest(object):
         if datasource_item.owner_id:
             owner_element = ET.SubElement(datasource_element, 'owner')
             owner_element.attrib['id'] = datasource_item.owner_id
+
+        datasource_element.attrib['isCertified'] = str(datasource_item.certified).lower()
+
+        if datasource_item.certification_note:
+            datasource_element.attrib['certificationNote'] = str(datasource_item.certification_note)
+
         return ET.tostring(xml_request)
 
     def publish_req(self, datasource_item, filename, file_contents, connection_credentials=None):
@@ -141,6 +156,8 @@ class ProjectRequest(object):
             project_element.attrib['description'] = project_item.description
         if project_item.content_permissions:
             project_element.attrib['contentPermissions'] = project_item.content_permissions
+        if project_item.parent_id:
+            project_element.attrib['parentId'] = project_item.parent_id
         return ET.tostring(xml_request)
 
 
@@ -213,6 +230,10 @@ class SiteRequest(object):
             site_element.attrib['disableSubscriptions'] = str(site_item.disable_subscriptions).lower()
         if site_item.subscribe_others_enabled:
             site_element.attrib['subscribeOthersEnabled'] = str(site_item.subscribe_others_enabled).lower()
+        if site_item.revision_limit:
+            site_element.attrib['revisionLimit'] = str(site_item.revision_limit)
+        if site_item.subscribe_others_enabled:
+            site_element.attrib['revisionHistoryEnabled'] = str(site_item.revision_history_enabled).lower()
         return ET.tostring(xml_request)
 
     def create_req(self, site_item):
@@ -331,6 +352,13 @@ class WorkbookConnection(object):
         return ET.tostring(xml_request)
 
 
+class TaskRequest(object):
+    @_tsrequest_wrapped
+    def run_req(xml_request, task_item):
+        # Send an empty tsRequest
+        pass
+
+
 class RequestFactory(object):
     Auth = AuthRequest()
     Datasource = DatasourceRequest()
@@ -341,6 +369,7 @@ class RequestFactory(object):
     Schedule = ScheduleRequest()
     Site = SiteRequest()
     Tag = TagRequest()
+    Task = TaskRequest()
     User = UserRequest()
     Workbook = WorkbookRequest()
     WorkbookConnection = WorkbookConnection()