+++ tableauserverclient/server/server.py
@@ -1,8 +1,9 @@
 import xml.etree.ElementTree as ET
 
 from .exceptions import NotSignedInError
+from ..namespace import Namespace
 from .endpoint import Sites, Views, Users, Groups, Workbooks, Datasources, Projects, Auth, \
-    Schedules, ServerInfo, ServerInfoEndpointNotFoundError
+    Schedules, ServerInfo, Tasks, ServerInfoEndpointNotFoundError
 
 import requests
 
@@ -21,7 +22,7 @@ class Server(object):
         Overwrite = 'Overwrite'
         CreateNew = 'CreateNew'
 
-    def __init__(self, server_address):
+    def __init__(self, server_address, use_server_version=False):
         self._server_address = server_address
         self._auth_token = None
         self._site_id = None
@@ -40,6 +41,11 @@ class Server(object):
         self.projects = Projects(self)
         self.schedules = Schedules(self)
         self.server_info = ServerInfo(self)
+        self.tasks = Tasks(self)
+        self._namespace = Namespace()
+
+        if use_server_version:
+            self.use_server_version()
 
     def add_http_options(self, options_dict):
         self._http_options.update(options_dict)
@@ -78,12 +84,21 @@ class Server(object):
 
         return version
 
-    def use_highest_version(self):
+    def use_server_version(self):
         self.version = self._determine_highest_version()
+
+    def use_highest_version(self):
+        self.use_server_version()
+        import warnings
+        warnings.warn("use use_server_version instead", DeprecationWarning)
 
     @property
     def baseurl(self):
         return "{0}/api/{1}".format(self._server_address, str(self.version))
+
+    @property
+    def namespace(self):
+        return self._namespace()
 
     @property
     def auth_token(self):