+++ test/test_site.py
@@ -91,7 +91,7 @@ class SiteTests(unittest.TestCase):
             single_site = TSC.SiteItem(name='Tableau', content_url='tableau',
                                        admin_mode=TSC.SiteItem.AdminMode.ContentAndUsers,
                                        user_quota=15, storage_quota=1000,
-                                       disable_subscriptions=True)
+                                       disable_subscriptions=True, revision_history_enabled=False)
             single_site._id = '6b7179ba-b82b-4f0f-91ed-812074ac5da6'
             single_site = self.server.sites.update(single_site)
 
@@ -100,6 +100,8 @@ class SiteTests(unittest.TestCase):
         self.assertEqual('Suspended', single_site.state)
         self.assertEqual('Tableau', single_site.name)
         self.assertEqual('ContentAndUsers', single_site.admin_mode)
+        self.assertEqual(True, single_site.revision_history_enabled)
+        self.assertEqual(13, single_site.revision_limit)
         self.assertEqual(True, single_site.disable_subscriptions)
         self.assertEqual(15, single_site.user_quota)